adminpanel/requirements.txt
Jinja2==3.0.3
pika==1.2.1
pydantic==1.9.0
psycopg2-binary==2.9.2
django-environ==0.8.1
Django>=3.0,<4.0